requirements.txt
asgiref==3.8.1
Django==5.0.4
django-appconf==1.0.6
django-cors-headers==4.3.1
django-isbn-field==0.5.3
django-select2==8.1.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
python-stdnum==1.20
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
